[Fix/#7] CalendarCollectionViewCell로 이름 수정
